feat: implement ChartContainer with BarChartCondition and StatusTrendChart components
